Upload files

+++ wp-content/plugins/duplicator-pro/classes/entities/class.storage.entity.php
@@ -131,7 +131,9 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
     public $s3_access_key;
     public $s3_bucket;
     public $s3_max_files                 = 10;
-    public $s3_region;
+    public $s3_provider                  = 'amazon';
+    public $s3_region                    = '';
+    public $s3_endpoint                  = '';
     public $s3_secret_key;
     public $s3_storage_class             = 'STANDARD';
     public $s3_storage_folder            = '';
@@ -190,6 +192,13 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
         $instance->onedrive_token_obtained      = $storage_data->onedrive_token_obtained;
         $instance->onedrive_authorization_state = $storage_data->onedrive_authorization_state;
 
+        //ONEDRIVE BUSINESS FIELDS
+
+        $instance->onedrive_endpoint_url        = isset($storage_data->onedrive_endpoint_url) ? $storage_data->onedrive_endpoint_url : '';
+        $instance->onedrive_resource_id         = isset($storage_data->onedrive_resource_id) ? $storage_data->onedrive_resource_id : '';
+        $instance->onedrive_storage_folder_id   = isset($storage_data->onedrive_storage_folder_id) ? $storage_data->onedrive_storage_folder_id : '';
+        $instance->onedrive_storage_folder_web_url = isset($storage_data->onedrive_storage_folder_web_url) ? $storage_data->onedrive_storage_folder_web_url : '';
+
         // FTP FIELDS
         $instance->ftp_server          = $storage_data->ftp_server;
         $instance->ftp_port            = $storage_data->ftp_port;
@@ -224,7 +233,9 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
         $instance->s3_access_key     = $storage_data->s3_access_key;
         $instance->s3_bucket         = $storage_data->s3_bucket;
         $instance->s3_max_files      = $storage_data->s3_max_files;
+        $instance->s3_provider       = $storage_data->s3_provider;
         $instance->s3_region         = $storage_data->s3_region;
+        $instance->s3_endpoint       = $storage_data->s3_endpoint;
         $instance->s3_secret_key     = $storage_data->s3_secret_key;
         $instance->s3_storage_class  = $storage_data->s3_storage_class;
         $instance->s3_storage_folder = $storage_data->s3_storage_folder;
@@ -342,7 +353,9 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
         $this->s3_bucket         = $source_storage->s3_bucket;
         $this->s3_access_key     = $source_storage->s3_access_key;
         $this->s3_secret_key     = $source_storage->s3_secret_key;
+        $this->s3_provider       = $source_storage->s3_provider;
         $this->s3_region         = $source_storage->s3_region;
+        $this->s3_endpoint       = $source_storage->s3_endpoint;
         $this->s3_storage_class  = $source_storage->s3_storage_class;
 
         //ONEDRIVE FIELDS
@@ -402,6 +415,14 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
 
     public function onedrive_is_business(){
         return !empty( $this->onedrive_endpoint_url) && !empty($this->onedrive_resource_id);
+    }
+
+    public function s3_is_amazon() {
+        if ((empty($this->s3_provider) || (!empty($this->s3_provider) && 'amazon' == $this->s3_provider))) {
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public function get_onedrive_storage_folder(){
@@ -680,7 +701,7 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
 
     public function get_full_s3_client()
     {
-        return DUP_PRO_S3_U::get_s3_client($this->s3_region, $this->s3_access_key, $this->s3_secret_key);
+        return DUP_PRO_S3_U::get_s3_client($this->s3_region, $this->s3_access_key, $this->s3_secret_key, $this->s3_endpoint);
     }
 
     public static function delete_by_id($storage_id)
@@ -720,6 +741,13 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
         }
 
         return $storage;
+    }
+
+    public static function is_exist($id) {
+        if (DUP_PRO_Virtual_Storage_IDs::Default_Local == $id) {
+            return true;
+        }
+        return self::is_exist_by_id_and_type($id, get_class());
     }
 
     public function get_dropbox_combined_access_token($use_curl = true)
@@ -921,7 +949,7 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
                 break;
 
             case DUP_PRO_Storage_Types::S3:
-                $type_text = DUP_PRO_U::__('Amazon S3');
+                $type_text = $this->s3_is_amazon() ? DUP_PRO_U::__('Amazon S3') : DUP_PRO_U::__('S3-Compatible (Generic)');
                 break;
             case DUP_PRO_Storage_Types::OneDrive:
                 $type_text = !$this->onedrive_is_business() ? DUP_PRO_U::__('OneDrive') : DUP_PRO_U::__('OneDrive (B)');
@@ -1291,7 +1319,7 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
                                         $this->purge_old_dropbox_packages($dropbox);
                                     }
                                 }else{
-                                    DUP_PRO_LOG::traceError("Uploaded File is corrupted, the hashes don't match!");
+                                    DUP_PRO_LOG::traceError("Uploaded File is corrupted, the hashes don't match! Is account out of space?");
                                     $upload_info->increase_failure_count();
 
                                 }
@@ -1779,9 +1807,18 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
     public function get_storage_location_string()
     {
         switch ($this->storage_type) {
-            case DUP_PRO_Storage_Types::Dropbox:
-                return "https://dropbox.com/home/Apps/Duplicator%20Pro/$this->dropbox_storage_folder";
+            case DUP_PRO_Storage_Types::Dropbox:                    
+
+                $dropbox = $this->get_dropbox_client();           
+                $dropBoxInfo = $dropbox->GetAccountInfo();
+               
+                if ($dropBoxInfo->locale == 'en') {
+                    return "https://dropbox.com/home/Apps/Duplicator%20Pro/$this->dropbox_storage_folder";
+                } else {
+                    return "https://dropbox.com/home";
+                }
                 break;
+            
 
             case DUP_PRO_Storage_Types::FTP:
                 return "ftp://$this->ftp_server:$this->ftp_port/$this->ftp_storage_folder";
@@ -2056,6 +2093,7 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
                 } catch (Exception $e) {
                     $upload_info->failed = true;
                     $upload_info->increase_failure_count();
+                    error_log("Problems copying package $package->Name to $this->sftp_storage_folder. " + $e->getMessage());
                     DUP_PRO_LOG::traceError("Problems copying package $package->Name to $this->sftp_storage_folder. " + $e->getMessage());
                 }
             }else{
@@ -2125,7 +2163,7 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
 
             if ($this->local_max_files > 0) {
 
-                if ($this->purge_package_record) {
+                if (isset($this->purge_package_record) && $this->purge_package_record) {
                     global $wpdb;
                     $table = $wpdb->base_prefix . "duplicator_pro_packages";
                     $purge_name_hash_arr = array();   
@@ -2170,7 +2208,7 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
                     $dirs_txt_filepath  = str_replace($suffix, '_dirs.txt', $archive_filepath);
 
                     
-                    if ($this->purge_package_record) {
+                    if (isset($this->purge_package_record) && $this->purge_package_record) {
                         $purge_name_hash_arr[] = $wpdb->prepare("%s", basename($archive_filepath, $suffix));
                     }
                                         
@@ -2185,7 +2223,7 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
                 }
                 
                 // Purge package record logic
-                if ($this->purge_package_record && !empty($purge_name_hash_arr)) {
+                if (isset($this->purge_package_record) && $this->purge_package_record && !empty($purge_name_hash_arr)) {
                     $max_created = $wpdb->get_var("SELECT max(created) FROM ".$table." WHERE concat_ws('_', name, hash) IN (".implode(', ', $purge_name_hash_arr).")");
                     $sql = $wpdb->prepare("DELETE FROM ".$table." WHERE created <= %s AND status = %d", $max_created, 100);
                     $wpdb->query($sql);
@@ -2704,7 +2742,7 @@ class DUP_PRO_Storage_Entity extends DUP_PRO_JSON_Entity_Base
                 return DUP_PRO_U::__('Local');
 
             case DUP_PRO_Storage_Types::S3:
-                return DUP_PRO_U::__('Amazon S3');
+                return $this->s3_is_amazon() ? DUP_PRO_U::__('Amazon S3') : DUP_PRO_U::__('S3-Compatible (Generic)');
             case DUP_PRO_Storage_Types::OneDrive:
                 return !$this->onedrive_is_business() ? DUP_PRO_U::__('OneDrive') : DUP_PRO_U::__('OneDrive (B)');
